+++ src/actions/index.js
@@ -1,6 +1,7 @@
 import axios from 'axios';
 
 export const FETCH_FLASHCARDS = 'fetch_flashcards';
+export const FETCH_USER = 'fetch_user';
 
 const ROOT_URL = 'http://localhost:8080';
 
@@ -12,3 +13,11 @@ export const fetchCards = () => {
     payload: request
   };
 }
+
+export const fetchUser = () => async dispatch => {
+  const res = await axios.get('/api/current_user');
+
+  dispatch({ type: FETCH_USER, payload: res.data});
+};
+
+
